feat: added controller and service to batch

+++ src/main/java/com/mundim/ticketbackendspringboot/dto/response/BatchResponseDto.java
@@ -11,7 +11,7 @@ import java.time.LocalDateTime;
 @ToString
 public class BatchResponseDto {
     private Long id;
-    private int percentage;
+    private Double percentage;
     private LocalDateTime batchStopDate;
     private int salesQtd;
     private String description;
